Ensure substr() receives length as integer

+++ lib/ASN1/Type/Primitive/ObjectIdentifier.php
@@ -72,7 +72,7 @@ class ObjectIdentifier extends Element
         int &$offset): ElementBase
     {
         $idx = $offset;
-        $len = Length::expectFromDER($data, $idx)->length();
+        $len = (int) Length::expectFromDER($data, $idx)->length();
         $subids = self::_decodeSubIDs(substr($data, $idx, $len));
         $idx += $len;
         // decode first subidentifier according to spec section 8.19.4
